+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.0.1
 * Define `Castable` type class
 * Define `cast` method
-* Define `I` data type
+* Define `Identity` data type
 * Define `Supertype` type family
 * Define `U_I_I` data type
 * Define `U_I_II` data type